Support for Int64 in SchemaBuilder

+++ src/NzbDrone.Core/Notifications/PushBullet/TestPushBulletCommand.cs
@@ -13,6 +13,6 @@ namespace NzbDrone.Core.Notifications.PushBullet
             }
         }
         public string ApiKey { get; set; }
-        public int DeviceId { get; set; }
+        public long DeviceId { get; set; }
     }
 }
